+++ index.js
@@ -11,7 +11,7 @@ import { askMasked, printShortcutHint, ORANGE, RED, RESET, BOLD, DIM } from "./s
 import { existsSync, readdirSync, unlinkSync } from "fs";
 import { dirname, join } from "path";
 
-const VERSION = "0.9.7";
+const VERSION = "0.9.8";
 
 // Clean up leftover files from previous /update (old EXEs that couldn't be deleted while running)
 function startupCleanup() {